+++ qiskit_experiments/library/characterization/analysis/__init__.py
@@ -17,7 +17,7 @@ from .fine_half_angle_analysis import FineHalfAngleAnalysis
 from .fine_amplitude_analysis import FineAmplitudeAnalysis
 from .fine_drag_analysis import FineDragAnalysis
 from .fine_frequency_analysis import FineFrequencyAnalysis
-from .remsey_xy_analysis import RamseyXYAnalysis
+from .ramsey_xy_analysis import RamseyXYAnalysis
 from .t2ramsey_analysis import T2RamseyAnalysis
 from .t1_analysis import T1Analysis
 from .cr_hamiltonian_analysis import CrossResonanceHamiltonianAnalysis
